Move flags parsing to its own file
